added tests and moved scheduler
